+++ myems-api/reports/enterproduction.py
@@ -1,11 +1,9 @@
-import re
 from datetime import datetime, timedelta, timezone
 from decimal import Decimal
 import falcon
 import mysql.connector
 import simplejson as json
 import config
-from core import utilities
 from core.useractivity import access_control, api_key_control